README-hacking: updating maneage branch moved to new section
SUMMARY: this commit does not affect your project.

Until this commit, the checklist to update the Maneage branch of a project
was within a separate subsection and the descriptions where comments in a
large code block. This made it hard to find and follow in the large
README-hacking.md file. Also, I noticed that the 'tarballs-prepare.sh'
didn't have a copyright notice and that one of the demo copyright
statements in README-hacking was wrong.

With this commit, the large code block has been broken into an enumerated
list (with small code-boxes under each item) and brought into a dedicated
section, which makes it easier to find. The two other minor points have
also been addressed.