Configuration: typo corrections in configure.sh
SUMMARY: this is just a houskeeping commit (no effect on your project).

Until this commit, two of the configure.sh checks for being on a macOS
system had typos (and were thus being wrongly evaluated). In one case, the
variable was called with a '$$' (instead of a single '$') and in the other
a '!=' was used (instead of a '=') resulting in a macOS command ('otool')
being called on GNU/Linux systems.

With this commit, both typos are corrected.