Configuration: more portable /bin/sh derivation of linked libraries
Summary: this will not affect the analysis of any project.

Until this commit, the part of the configuration script which would check
for the necessary host libraries used by the host shell was not portable
(would not work on some systems). Also, the list of paths had an extra ':'
at the end (allowing the searching of the running directory for libraries,
which is a bug and can have unpredictable effects).

With this commit, both issues have been addressed in the configuration
script.