Configuration: only default software in TARGETS.conf is Gnuastro
Summary: this will not affect your high-level projects

Until this commit, the 'reproduce/software/config/TARGETS.conf' included
both Gnuastro and Astropy. But the default analysis does not use Astropy,
therefore to avoid taking too much time for a default build, it should only
be added when the user needs it. It was present because of the tests in the
previous commit to update the Python environment, but I forgot to remove it
before pushing.

With this commit, TARGETS.conf only contains Gnuastro.