Configuration: Python packages can be reinstalled without crash
Summary: this will not affect any analysis.

Until this commit, there were often fatal errors when re-installing a
python package in the Maneage system (after events such as: an error during
the install; a computer reboot; a modification of dependencies or build
rules in 'high-level.mk'; or a removal of the successful-install indicator
file '.local/version-info/python/PACKAGENAME-PACKAGEVERSION'). The reason
for this particular failure was that 'python-installer' considers file
clobbering - replacing an old existing file by a new file - to be a fatal
error; the python function 'FileExistsError' is used to report the error
and terminate execution.

With this commit, the build rule for 'python-installer' comments out the
'FileExistsError' line in 'src/installer/destinations.py'. Based on several
tests of interrupted installs of python packages, it appears that this
commit solves this bug, allowing convenient re-installs of python packages.