Configuration: basic software build with host GCC 15.1
Summary: this will not affect already built/configured projects, only
improving the portability of future builds.

Until this commit, no specific C standard version was given during the
build of Basic software (which use the host's c compiler, culminating in
the building of GCC within Maneage). On the other hand, GCC 15.1 was
recentely released and made available on some operating systems. GCC 15.1
has made C23 its standard C version (from C17), as a result, some of the
basic software crashed and didn't allow Maneage to built on those operating
systems.

With this commit, the '-std=c17' flag has been added to software that
crashed when the host GCC was 15.1. Also, they have been grouped with a
description of this problem in 'versions.conf' so we check the need for
this option in future version updates.

In parallel, some minor edits/clarifications were made in the two
'README-apptainer.md' and 'README-docker.md' files to be more useful.