make dist: better comments within the dist rule of initialize
Summary: will not affect existing configuration or analysis.

Until now, some of the comments in the 'dist' rule of initialize (to build
the tarball of the project) were out of date (for example referenced the
'LOCAL.conf' file in the top source directory) or not sufficiently clear.

With this commit, those comments have been edited to improve clarity.