PDF building: pdflatex uses local shell's LD_LIBRARY_PATH
Summary: will not affect existing configuration or analysis.

Until this commit, if the Maneage readline library (used by the shell) did
not match the local readline library, the 'pdflatex' command would crash
when building the template Maneage branch. This was because PGFPlots uses
the local shell and we do not build TeXLive from source within Maneage
yet. Previously, we had added a modification to 'LD_LIBRARY_PATH' before
the 'pdflatex' command to use 'sys_library_sh_path', but the variable was
never initilized (because it was only available within the configuration
phase)!

With this commit, the problem is solved by adding that variable within
'LOCAL.conf' so it is available in the analysis phase of a Maneage'd
project also. In order to do so, the set of commands that set the
'sys_library_sh_path' variable within 'configure.sh' was brought up before
the creation of 'LOCAL.conf'.

In case you have had this problem, the solution is to delete
'.build/software/config/LOCAL.conf' and re-configure your project by
manually setting the necessary directories. In this way, the 'LOCAL.conf'
is created again with the new variable.

This bug was found with the help of Faezeh Bidjarchian.